создает моки для развернутого offer
