+++ src/mock/offers-preview.ts
@@ -1,6 +1,6 @@
-import { Offer } from '../types/offers';
+import { OfferPreview } from '../types/offers-preview';
 
-export const offer: Offer[] = [
+export const offersPreview: OfferPreview[] = [
   {
     id: '6af6f711-c28d-4121-82cd-e0b462a27f00',
     title: 'Beautiful & luxurious studio at great location',
@@ -22,21 +22,7 @@ export const offer: Offer[] = [
     isFavorite: false,
     isPremium: false,
     rating: 3,
-    previewImage: 'https://url-to-image/image.png',
-    description: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    bedrooms: 3,
-    goods: [
-      'Heating'
-    ],
-    host:{
-      name:'Oliver Conner',
-      avatarUrl:'https://url-to-image/image.png',
-      isPro: false,
-    },
-    images: [
-      'https://url-to-image/image.png'
-    ],
-    maxAdults: 4,
+    previewImage: 'https://url-to-image/image.png'
   },
   {
     id: '9af6f221-h28f-3521-82cd-e0b882a27f00',
@@ -59,22 +45,7 @@ export const offer: Offer[] = [
     isFavorite: false,
     isPremium: false,
     rating: 5,
-    previewImage: 'https://url-to-image/image.png',
-    description: 'A big house near Elbphilharmonie.',
-    bedrooms: 5,
-    goods: [
-      'Heating',
-      'Veranda'
-    ],
-    host:{
-      name:'Filip Snow',
-      avatarUrl:'https://url-to-image/image.png',
-      isPro: false,
-    },
-    images: [
-      'https://url-to-image/image.png'
-    ],
-    maxAdults: 6,
+    previewImage: 'https://url-to-image/image.png'
   },
   {
     id: '4gf2f449-h99k-1121-82lm-e0b882a55f00',
@@ -97,23 +68,7 @@ export const offer: Offer[] = [
     isFavorite: false,
     isPremium: true,
     rating: 4,
-    previewImage: 'https://url-to-image/image.png',
-    description: 'Beautiful small flat in the heart of the city.',
-    bedrooms: 2,
-    goods: [
-      'Heating',
-      'Breakfast'
-    ],
-    host:{
-      name:'Lolita Milyavskaya',
-      avatarUrl:'https://url-to-image/image.png',
-      isPro: true,
-    },
-    images: [
-      'https://url-to-image/image.png',
-      'https://url-to-image/image.png'
-    ],
-    maxAdults: 3,
+    previewImage: 'https://url-to-image/image.png'
   },
   {
     id: '7ff2d234-s99s-2222-97lg-e0b882a55f00',
@@ -134,24 +89,8 @@ export const offer: Offer[] = [
       zoom: 8
     },
     isFavorite: true,
-    isPremium: true,
+    isPremium: false,
     rating: 5,
-    previewImage: 'https://url-to-image/image.png',
-    description: 'Modern mansion in a quiet area.',
-    bedrooms: 6,
-    goods: [
-      'Heating',
-      'Breakfast'
-    ],
-    host:{
-      name:'Alex Partini',
-      avatarUrl:'https://url-to-image/image.png',
-      isPro: true,
-    },
-    images: [
-      'https://url-to-image/image.png',
-      'https://url-to-image/image.png'
-    ],
-    maxAdults: 10,
+    previewImage: 'https://url-to-image/image.png'
   },
 ];
