+++ src/types/offers.ts
@@ -1,23 +1,16 @@
-export type Location = {
-  latitude: number;
-  longtitude:number;
-  zoom:number;
-}
+import {OfferPreview } from './offers-preview';
 
-export type City = {
-  name:string;
-  location:Location;
-}
 
-export type Offer = {
-  id: string;
-  title: string;
-  type: string;
-  price: number;
-  city: City;
-  location: Location;
-  isFavorite: boolean;
-  isPremium : boolean;
-  rating: number;
-  previewImage: string;
+export type Host = {
+  name: string;
+  avatarUrl: string;
+  isPro: boolean;
+}
+export type Offer = OfferPreview & {
+  description: string;
+  bedrooms: number;
+  goods: string[];
+  host: Host;
+  images: string[];
+  maxAdults: number;
 }
